--- server/config/authMiddleware.js
@@ -1,9 +1,0 @@
-const dotenv = require('dotenv');
-dotenv.load();
-
-const jwt = require('express-jwt');
-
-module.exports = jwt({
-    secret: new Buffer(process.env.AUTH0_CLIENT_SECRET, 'base64'),
-    audience: process.env.AUTH0_CLIENT_ID
-});
