--- server/config/userInViews.js
@@ -1,11 +1,0 @@
-/**
- * The purpose of this middleware is to have the `user`
- * object available for all views.
- *
- */
-module.exports = function () {
-    return function (req, res, next) {
-      res.locals.user = req.user;
-      next();
-    };
-  };
