+++ server/package.json
@@ -20,6 +20,7 @@
 		"testEnvironment": "node"
 	},
 	"dependencies": {
+		"connect-ensure-login": "^0.1.1",
 		"cookie-parser": "^1.4.3",
 		"cors": "^2.8.5",
 		"dotenv": "^6.1.0",
