--- server/routes/oldauth/auth.js
@@ -1,32 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const passport = require('passport');
-
-// Perform the login, after login Auth0 will redirect to callback
-router.get('/signin', passport.authenticate('auth0', {
-	scope: 'openid email profile'
-  }), function (req, res) {
-	res.redirect('/');
-  });
-  
-  // Perform the final stage of authentication and redirect to previously requested URL or '/user'
-  router.get('/callback', function (req, res, next) {
-	passport.authenticate('auth0', function (err, user, info) {
-	  if (err) { return next(err); }
-	  if (!user) { return res.redirect('/signin'); }
-	  req.logIn(user, function (err) {
-		if (err) { return next(err); }
-		const returnTo = req.session.returnTo;
-		delete req.session.returnTo;
-		res.redirect(returnTo || '/user');
-	  });
-	})(req, res, next);
-	});
-  
-  // Perform session logout and redirect to homepage
-  router.get('/signout', (req, res) => {
-	req.logout();
-	res.redirect('/');
-  });
-
-module.exports = router;
